renamed `register_services.js` to `register_routes.js`
